requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-filter==22.1
django-phonenumber-field==7.0.0
django-twilio==0.14.3
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
phonenumbers==8.12.57
PyJWT==2.6.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
twilio==7.15.1
urllib3==1.26.12
zipp==3.9.0
